error checks wip

+++ be/app.py
@@ -1,24 +1,30 @@
-from flask import Flask, jsonify, request, json
+from flask import Flask, jsonify, request, json, abort
 from flask_mysqldb import MySQL
 from flask_cors import CORS
+import os 
 
 app = Flask(__name__)
 
 app.config.from_pyfile('settings.py')
 
 mysql = MySQL(app)
-CORS(app, origins=["http://localhost:3000"])
+
+CORS(app, origins=[os.environ.get("CLIENT_URL")])
 
 #students
 @app.route("/students/<id>", methods=["DELETE"])
 def delete_students(id):
     if request.method == "DELETE":
+        if student_res is None:
+            abort(400, "Insufficient fields")
+
         query = f"DELETE FROM student WHERE id={id}"
 
         cursor = mysql.connection.cursor()
         student_res = cursor.execute(query)
-        print("RESPONSE", student_res)
-        #checks
+        if student_res == 0:
+            abort(400, "Not Deleted")
+
         student_res = mysql.connection.commit()
         cursor.close()
 
@@ -47,11 +53,15 @@ def students():
         last_name = data['last_name']
         email = data["email"]
         dob = data['dob']
+        if first_name is None or last_name is None or email is None or dob is None:
+            abort(400, "Insufficient fields")
 
         query = f"INSERT INTO student (first_name, last_name, dob, email) VALUES('{first_name}', '{last_name}', '{dob}', '{email}');"
         cursor = mysql.connection.cursor()
         student_res = cursor.execute(query)
-        #checks
+        if student_res == 0:
+            abort(400, "Not Created")
+
         student_res = mysql.connection.commit()
         cursor.close()
 
@@ -77,12 +87,16 @@ def courses():
     if request.method == "POST":
         data = json.loads(request.data)
         name = data['name']
+        if name is None:
+            abort(400, "Insufficient fields")
 
         query = f"INSERT INTO course (name) VALUES('{name}')"
 
         cursor = mysql.connection.cursor()
         course_res = cursor.execute(query)
-        #checks
+        if course_res == 0:
+            abort(400, "Not Created")
+
         course_res = mysql.connection.commit()
         cursor.close()
 
@@ -94,10 +108,15 @@ def courses():
 @app.route("/courses/<id>", methods=["DELETE"])
 def delete_courses(id):
     if request.method == "DELETE":
+        if id is None:
+            abort(400, "Insufficient fields")
+
         query = f"DELETE FROM course WHERE id={id}"
 
         cursor = mysql.connection.cursor()
         course_res = cursor.execute(query)
+        if course_res == 0:
+            abort(400, "Not Deleted")
         #checks
         course_res = mysql.connection.commit()
         cursor.close()
@@ -115,12 +134,16 @@ def results():
         student = data['student']
         course = data["course"]
         score = data["score"]
-        
+        if student is None or data is None or course is None or score is None:
+            abort(400, "Insufficient fields")
+
         query = f"INSERT INTO result (student, course, score) VALUES('{student}', '{course}', '{score}')"
 
         cursor = mysql.connection.cursor()
         results_res = cursor.execute(query)
-        #checks
+        if results_res == 0:
+            abort(400, "Not Created")
+
         results_res = mysql.connection.commit()
         cursor.close()
 
@@ -133,7 +156,6 @@ def results():
         results = db.execute(query)
 
         results = db.fetchall()
-        print("RESULTS", results)
 
         db.close()
 
@@ -145,11 +167,15 @@ def results():
 @app.route("/results/<id>", methods=["DELETE"])
 def delete_results(id):
     if request.method == "DELETE":
+        if id is None:
+            abort(400, "Insufficient fields")
         query = f"DELETE FROM result WHERE id={id}"
 
         cursor = mysql.connection.cursor()
         result_res = cursor.execute(query)
-        #checks
+        if result_res == 0:
+            abort(400, "Not Deleted")
+
         result_res = mysql.connection.commit()
         cursor.close()
